Moved code in framework/hooks to framework/internal/hooks
