+++ frontend/src/framework/internal/hooks/moduleHooks.ts
